RelationshipDB.get(relationship) based adjustments
relationship sync on CCUD not ready yet!
